feat: udpate kuzu wasm version
