+++ examples/graphy/src/pages/components/index.tsx
@@ -2,5 +2,3 @@ export { default as CreateHeaderPortal } from './CreateHeaderPortal';
 export { default as GraphList } from './GraphList';
 export { default as GraphSchema } from './GraphSchema';
 export { default as Container } from './Container';
-/** app */
-export { default as ClusterGraph } from './ClusterGraph';
